Introduce DownloaderParams in AiCoreFactory
This CL introduces a new mojom param when creating
`ModelDownloaderAndroid`. This param is defined as a mojom struct to
facilitate convenient message passing between C++ and Java.

Currently the only field added in this param is
`only_download_in_persistent_mode`. This is based on the feedback from
Android System Health Council that certain Chrome features should only
be enabled when persistent mode is enabled. It is set to true by
default, but it is worth noting that using the default value will limit
features to a small set of devices.

A future CL will add a feature flag to set
`only_download_in_persistent_mode` to true.

Internal CL: http://cl/815909259

Bug: 449213567
Change-Id: I76fe8f26eaaa647507593b0d49a7ca7f23401852
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7014037
Reviewed-by: Steven Holte <holte@chromium.org>
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Siddhartha S <ssid@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1529755}