Add workaround to ensure WaitSyncToken() for dummy WebGPU Texture
When canvas resource copying involves a dummy WebGPU Texture, the
actual copying is skipped but we still need to wait on the underlying
resource. This waiting is already ensured in the current code, but
with our upcoming change of deferring SyncToken waiting, this guarantee
will no longer be automatically satisfied. Therefore we introduce a
workaround in this CL to add back the guarantee of WaitSyncToken()
for dummy WebGPU Textures.

Bug: 40286368
Change-Id: I813f068b772d96c731ff6d2815436d6fe08655a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7027971
Commit-Queue: Mingjing Zhang <mjzhang@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1529759}