[soft navs] Add trace events for scroll restoration debugging
This CL adds a a trace event to PaintTimingDetector::NotifyScroll with
the scroll reason and to SoftNavigationContext::OnInputOrScroll with the
painted area to help debug scroll restoration use cases. We're ignoring
programmatic scrolls, so we shouldn't be stopping the heuristics for
JS-driven scroll restoration, but something might be broken. These
events should help root cause suspected scroll restoration failures,
or help determine if it's something else.

Bug: 330709851
Change-Id: I9560bfe9500acb5907a9d347fc4fc37602483036
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7038081
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1529852}