[Tab] Allow mPendingLoadParams for tabs with WebContents
For tab group sync we want to support updates to background tabs without
triggering a navigation until the tab is selected.

Historically, we achieved this by freezing a tab's WebContents to a
WebContentsState and appending a serialized navigation entry. This
works, but it destroys the WebContents. We are working on changes to
ensure that post startup every tab has a 1:1 relationship with its
WebContents.

To address this, this CL adopts a strategy similar to DeferredTabState
from desktop where the title and URL are updated in the UI, but the
navigation is deferred until the WebContents is next activated.
Meanwhile, the current WebContents state is discarded to make it seem
like a navigation has sort of happened (media is paused, etc.).

The following cases need to be handled:

Case 1: Only mPendingLoadParams exists - clobber it with the new
LoadUrlParams.

Case 2: mWebContentsState exists - append a new pending navigation to
the end of the serialized navigation entries.

Case 3: mWebContents exists - set mPendingLoadParams to the new
LoadUrlParams clobbering the previous entry if it existed.

For case 1 and case 3, there is a change from the current approach as
there only be one pending navigation at a time. I.e. if a Tab receives
another pending navigation before the current mPendingLoadUrlParams is
consumed the previous pending load will get clobbered rather than being
added to the navigation history of the tab.

Without further modification, case 2 was inconsistent with cases 1 & 3
as it allowed unlimited pending load params to be stacked. However, this
could be changed to be consistent with cases 1 & 3 in a few ways;

1) Appending could be tracked in WebContentsState such as to only append
a single entry.

2) We could create a WebContents from the WebContentsState then set a
mPendingLoadParams. This works, but does extra work.

3) Allow an mPendingLoadParams with an mWebContentsState. This makes the
implementation of `loadIfNeeded()` more complicated and significantly
complicates TabStateExtractor as well.

4) Get rid of case 2 by always having a WebContents.

5) Ignore the problem since it is somewhat niche in scope.

I've chosen to go with option 1 as it solved the problem with the least
difficulty.

With both flags enabled this should entirely eliminate cases where
WebContents get destroyed in TabImpl. That said, testing and
verification is required to ensure this doesn't introduce any
regressions as a mPendingLoadParams + mWebContents is a previously
untested configuration.

Bug: 449784092
Change-Id: I229d442b2dfb53780e0903638d50763213dfe4f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7040224
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1529859}