win: Add opaque background for underlay videos
This change forces a solid black background on all underlay videos to
avoid seeing through the video hole punch.

This fixes a scenario that can happen with Media Foundation videos. We
can end up with a valid MF surface proxy handle but have it not yet
ready on the MF renderer process side. In this case it will display
transparent black, allowing the user to see through th video hole punch.

We previously attempted to fix this MF case in crrev.com/c/6976796 but
that only addressed the interval between the first frame containing an
MF video and the subsequent frame with a valid swap chain handle in the
DCOMPSurfaceRegistry. It did not address the case described above.

Change-Id: I148ff7b7d78f1d2cffa4afe95a7e38a8a6751795
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7029419
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Michael Tang <tangm@microsoft.com>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1529868}