extensions: Construct ExtensionActionPlatformDelegateViews externally
This is a part of the patch series to move stateful UI logic from
ExtensionActionViewController to ExtensionActionPlatformDelegate.

This patch changes how ExtensionActionPlatformDelegate(Views) is
constructed. Formerly, the ExtensionActionViewController created and
owned the platform delegate, but it meant that we cannot pass
platform-specific arguments to the platform delegate's constructor.
With this patch, the owners of the view controller are responsible for
constructing the platform delegate and passing its ownership to the
view controller. The platform delegate's constructor now takes two
additional arguments, BrowserWindowInterface and ExtensionsContainer,
that we plan to use in subsequent patches.

This is a pure refactor. No functional changes are expected.

Bug: 448199168
Change-Id: Ibf93351212703910619fa0329d093fbd629c02e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7011827
Reviewed-by: Darryl James <dljames@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1529871}