[Happy Eyeballs v3] Remove excluded IP arg from MaybeAttemptTcpBased().
This seems of limited utility. There are 4 cases here:

1. There are other IP addresses not marked as slow (and not marked
as failing). In this case, we would try one of those other IP addresses
even without the exclusion, since the excluded IP address is now marked
as slow.

2. There is only one IP address not marked as failing, which is the
one that was just marked as slow. In that case, we wouldn't retry
that IP address, anyways, since we don't retry IPs marked as slow
if we already have enough connection attempts.

3. There are other IP addresses not marked as failing and not marked
as slow (which necessarily wouldn't marked the excluded one, which was
just marked as slow). We would use one of them, regardless of the
exclusion.

4. There is at least one IP address already marked as slow, and no
others marked as slow succeeded (or not marked as anything). In this
case, we would have started a connection to one of those other slow
IP addresses, in preference to the one just marked as slow. If there
are multiple IP addresses, for instance, once they've all been
marked as slow, I guess the exclusion means we'd potentially bounce
between making to the first two connections them, each time we mark
an attempt as slow. This seems a marked difference from 2, where
was just don't try to make any more connections.

Bug: 40812426
Change-Id: Id93fbde49b58f594202501241861d462634b5859
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7037754
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1529895}