disk_cache/sql: Treat disk full error as catastrophic in SqlDiskCache
The UMA metric "Net.SqlDiskCache.Backend.SqliteError" shows a
disproportionately high number of kFullDisk errors. This is because a
disk full error was not treated as a catastrophic error, causing it to
be logged repeatedly on every subsequent SQL operation.

While for critical databases, retrying might be a valid strategy, for
the HTTP Cache, the database is expendable. Deleting the database upon
a disk full error is a better approach as it frees up disk space and
prevents other errors.

This CL changes the behavior by treating SQLITE_FULL as a catastrophic
error within the Backend::DatabaseErrorCallback. This will trigger the
database to be razed, freeing up space and preventing the accumulation
of repetitive error logs.

Bug: 443171275
Change-Id: Ic48397795d4216b4f5faf18d058b730c2d9e7cbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7039803
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1529944}