net: Avoid PostTask() on HttpStreamPool switch for request
The HttpStreamFactory::JobController can be deleted when a request is
delegated to the HttpStreamPool. Previously, this was handled
asynchronously with `PostTask` to prevent use-after-free issues within
the controller's state machine (`DoLoop`). The scheduled task could have
delayed high priority tasks such as notifying created HttpStreams and
might cause some perf regression.

This change simplifies the lifetime management by making the deletion
synchronous. If the JobController is for request (not preconnect), the
`PostTask` is removed, and the controller is now deleted immediately.

To prevent the state machine from executing after the request is handed
over to the HttpStreamPool, `DoLoop` returns immediately and `RunLoop`
completes the JobController.

For preconnect, JobController still uses `PostTask`. We may consider
removing `PostTask` for preconnect as a follow-up.

Bug: 414173943
Change-Id: I6836763d70022ff0dc47b8cf574d8ffe62b46e41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7036899
Reviewed-by: mmenke <mmenke@chromium.org>
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1529952}