Refactor Android device connection states and add lock detection.
This change introduces a `ConnectedState` enum to better represent the
status of an Android device (e.g., connected, unauthorized, offline,
locked). The `adb devices` output is now parsed to distinguish between
these states. Additional ADB commands are run to detect if a connected
device is currently locked, and this information is propagated to the
DevTools UI. The UI messages are updated to provide more specific
feedback based on the device's state.

We no longer poll DevTools services on a device that is locked,
as this has caused issues with hanging the adb service.

Bug: 372055979
Change-Id: I54993462e261c3bfc3db5629af3ffb781b340128
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7037037
Auto-Submit: Eric Leese <leese@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1529963}