[YourSavedInfo] Refactor labels and icons for passwords category
This change introduces a new string for "Passkeys" and uses it for the
Passkeys chip on the "Your saved info" page. It also updates the icon
for the Passkeys chip to a dedicated passkey icon. The "Passwords" chip
label is also updated to use a localized string.

Result: https://screenshot.googleplex.com/6X5oLSwkXFa5MSt

Bug: 433612757
Change-Id: I853b41988cda96678edc9b9c540634c6ce8e2ab5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7039264
Reviewed-by: Stanislav Mikheyev <mikhe@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Julia Sobiech <jsobiech@google.com>
Cr-Commit-Position: refs/heads/main@{#1530017}