Implement HttpNoVarySearchData::CanonicalizeQuery()
Add CanonicalizeQuery() method to the HttpNoVarySearchData class. This
has the property that if two URLs have the same base URL and the return
value of CanonicalizeQuery() is the same, then AreEquivalent() would
return true for those URLs. This will permit NoVarySearchCache to
replace multiple calls to AreEquivalent() with a single call to
CanonicalizeQuery() and a hash lookup.

Also add a CreateFromRawHeaders() function to the HttpNoVarySearchData
unit tests to reduce the amount of repeated code in that file.

Bug: 403211337
Change-Id: I720f337323d0c0128f6fae2445bd084659dd59fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7039802
Commit-Queue: Adam Rice <ricea@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530029}