[//gpu] Remove kVaapiJpegImageDecodeAcceleration
We will not be shipping this feature. This CL does the minimal work to
rip it out; we will keep going on teardown in followup CLs.

Notes:

* This necessitates removing the ImageDecodeAcceleratorStub unittest,
as the tests there fundamentally rely on the feature being enabled (and
more deeply on the functionality being removed from
image_decode_accelerator_stub.cc in this CL). This is not a problem as
(a) with this CL we eliminate any path to ImageDecodeAcceleratorStub
actually being used, and (b) this CL chain is working toward removing
ImageDecodeAcceleratorStub itself in any case.
* Similarly, the VaapiImageDecodeAcceleratorWorker unittest now is also
no longer relevant as that class CHECK()'s in its constructor that it
has supported files (in production the static CREATE() method early-outs
when there are no supported profiles).
* Finally, the GpuImageDecodeCache accelerated decode tests are no
longer relevant.

Bug: 450466845
Change-Id: I3a94685fe56f40359c347a914ab2c9ba5fac9b48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7031820
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530039}