[iOS] Add FileUploadPanelMediator
This CL adds a FileUploadPanelMediator to link the file upload panel UI
to the model layer. In particular the file upload panel UI is tied to a
`ChooseFileController` instance which contains the parameters of the
file upload event and can be used to submit files to the web page.

Bug: 441659098
Change-Id: I1a28d07e0f38e3839d51c47eb901e0c5e715de0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7036753
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Auto-Submit: Quentin Pubert <qpubert@google.com>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Cr-Commit-Position: refs/heads/main@{#1530052}