[UNO] Support Shopping Collection with account storage for bookmarks
The shopping feature is currently only available for syncing users. With
the deprecation of `ConsentLevel::kSync`, and the migration of syncing
users to signed in users, this feature should work for the signed in
state.

With the launch of `SyncEnableBookmarksInTransportMode`, there is now a
separation between account bookmarks and local bookmarks. Therefore,
items for the shopping collection must be saved to the account rather
than locally.

This is done by choosing the account "Other bookmarks" node as the
default location for saving product bookmarks for which price tracking
should be enabled. If the user had already saved the bookmark locally, a
new bookmark is created in the account so that the feature works as
expected.

If the user is signed in, but does not have account nodes (e.g. disabled
by policy or syncing bookmarks was turned off in the account settings),
the feature is not available.

These changes are behind the flag `ReplaceSyncPromosWithSignInPromos`.

Bug: 352022179
Change-Id: I614840b1af78cf4d3a83cca5737086bc167ae20a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7003194
Auto-Submit: Amelie Schneider <amelies@google.com>
Reviewed-by: Mike Wasserman <msw@chromium.org>
Commit-Queue: Amelie Schneider <amelies@google.com>
Cr-Commit-Position: refs/heads/main@{#1530067}