[PermissionsAI] Rename IsAiv4ModelAvailable to IsAIv4FeatureEnabled
The function `IsAiv4ModelAvailable` was misleadingly named as it only checked for the `kPermissionsAIv4` feature flag and did not verify the actual availability of the prediction model.

This change renames the function to `IsAIv4FeatureEnabled` to more accurately reflect its purpose of checking the feature flag's status. This improves code clarity and maintainability by ensuring that the function's name aligns with its implementation.

Bug: 382447738
Change-Id: Ia20564326cf9325faaabb112313a4c92a1f04602
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7044430
Commit-Queue: Judith Hemp <hempjudith@google.com>
Reviewed-by: Elias Klim <elklm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530096}