net: Use priority task runners in HttpStreamPool
This change replaces the default task runner with a new
HttpStreamPool::TaskRunner() helper that returns a task runner based on
the request's priority.

This allows tasks within the `HttpStreamPool` and its related classes to
be scheduled according to the priority of the corresponding network
request. This prevents low-priority tasks delaying the completion of
high-priority tasks.

The HEv3 path is sometimes slow compared to the non-HEv3 path. One
hypothesis of the regression is the HEv3 path uses more PostTasks. This
change might mitigate high-priority tasks delays.

This functionality is controlled by the `kEnablePriorityTaskRunner`
feature parameter. The parameter is enabled by default as the priority
task runner is enabled by default now.

Bug: 414173943
Change-Id: I4cca8fce4b57284cef61d345ec6c1e2847443928
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7043710
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Hayato Ito <hayato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530097}