[iOS] Fix the corner radius applied to image in the share extension
When sharing an image, the preview in the share extension was displayed
with sharp, 90-degree corners instead of the intended rounded corners.
This occurred because the corner radius was applied directly to the
image view, but the image's content was not being clipped to the view's
rounded bounds. This CL fixes the issue by embedding the image view
within a new container view. The corner radius and clipsToBounds are set
on the container. The image view is set to fill the container using an
aspect fill content mode, which causes the image to be correctly
clipped, inheriting the container's rounded corners.

Before: https://screenshot.googleplex.com/9MXnRWJoRTMQriH
After: https://screenshot.googleplex.com/6iNjQbKx3KtqYin

Change-Id: I689b38e6be5f1632024b0155d7a9831135c93e66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7037891
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Elmehdi Rahmaoui <erahmaoui@google.com>
Cr-Commit-Position: refs/heads/main@{#1530110}