Make BrokeredUdpClientSocket stop making its own BrokerHelperWin.
Instead, it now uses the one from the BrokeredClientSocketFactory.
Each BrokerHelperWin gets a list of connected networks on
coonstruction, which according to our tracing profiler is fairly
expensive in practice.

The one downside of this is it relies on a single shared
BrokeredHelperWin updating its network list at the right time on
network changes, but we already use the BrokeredClientSocketFactory's
single BrokerHelperWin for all TCP socket brokering checks, so that's
not really a regression. I think we may well want to update the network
list every couple seconds, even without any network change, but that's
an issue for another day.

Fixed: 451644977
Change-Id: Ib6268b8c6f554186822a8061fa3ef19c6076bbde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7042502
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1530116}