[Refactor] Remove pointer to PrefService in ProfileUserData
SyncPromoIdentityPillManager used to hold a SigninPrefs instance which
internally stores a raw_ptr<PrefService> coming from the Profle.
However SyncPromoIdentityPillManager is held in a ProfileUserData.

This seems to cause issues with tests using
`BrowserWithTestWindowTest` as their base test suite, because of
destruction order; creating dangling pointer.

Fix: Hold a unique_ptr to SigninPrefs in order to clear it when needed.
Clearing it along with the IdentityManager shutting down, as it is
a notification that the Profile is shutting down. This is not a
common pattern, however there is no currently any notification that the
PrefService is shutting down.

Change-Id: I4e9f7e6cf27c7a69b63855b51dbd4edb0e3e16e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7037577
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1530230}